Client pipe should be provided by service provider
