+++ src/Api/ApiServiceProvider.php
@@ -102,6 +102,30 @@ class ApiServiceProvider extends AbstractServiceProvider
                 'discussionRenamed' => BasicDiscussionSerializer::class
             ];
         });
+
+        $this->container->singleton(Client::class, function($container) {
+            $pipe = new MiddlewarePipe;
+
+            $pipe->pipe(new Middleware\ResolveRouteFromName($container->make('flarum.api.routes')));
+
+            $middlewareStack = array_filter($container->make('flarum.api.middleware'), function ($middlewareClass) {
+                return !in_array($middlewareClass, [
+                    HttpMiddleware\ParseJsonBody::class,
+                    HttpMiddleware\StartSession::class,
+                    HttpMiddleware\AuthenticateWithSession::class,
+                    'flarum.api.route_resolver',
+                    HttpMiddleware\CheckCsrfToken::class
+                ]);
+            });
+
+            foreach ($middlewareStack as $middleware) {
+                $pipe->pipe($container->make($middleware));
+            }
+
+            $pipe->pipe(new HttpMiddleware\ExecuteRoute());
+
+            return new Client($pipe);
+        });
     }
 
     /**
