+++ src/Api/Client.php
@@ -18,6 +18,7 @@ use Flarum\User\User;
 use Illuminate\Contracts\Container\Container;
 use Laminas\Diactoros\ServerRequestFactory;
 use Laminas\Stratigility\MiddlewarePipe;
+use Laminas\Stratigility\MiddlewarePipeInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
@@ -26,65 +27,15 @@ use Psr\Http\Server\RequestHandlerInterface;
 class Client
 {
     /**
-     * @var MiddlewarePipe
+     * @var MiddlewarePipeInterface
      */
     protected $pipe;
 
     /**
-     * @var Registry
-     */
-    protected $registry;
-
-    /**
      * @param Container $container
-     * @param Registry $registry
      */
-    public function __construct(Container $container, Registry $registry)
+    public function __construct(MiddlewarePipeInterface $pipe)
     {
-        $this->registry = $registry;
-
-        $middlewareStack = $container->make('flarum.api.middleware');
-
-        $middlewareStack = array_filter($middlewareStack, function ($middlewareClass) {
-            return ! in_array($middlewareClass, [
-                HttpMiddleware\ParseJsonBody::class,
-                HttpMiddleware\StartSession::class,
-                HttpMiddleware\AuthenticateWithSession::class,
-                'flarum.api.route_resolver',
-                HttpMiddleware\CheckCsrfToken::class
-            ]);
-        });
-
-        $routeCollection = $container->make('flarum.api.routes');
-
-        $pipe = new MiddlewarePipe;
-        $pipe->pipe(new class($routeCollection) implements MiddlewareInterface {
-            public function __construct(RouteCollection $routeCollection)
-            {
-                $this->routeCollection = $routeCollection;
-            }
-
-            public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
-            {
-                $route = $this->routeCollection->getRoutes()[$request->getAttribute('routeName')];
-
-                $request = $request->withMethod($route['method']);
-                $request = $request->withAttribute('routeHandler', $route['handler']);
-                // These aren't available to be parsed out since we're passed
-                // a route name, not a URL. If needed, these can be specified
-                // as query params.
-                $request = $request->withAttribute('routeParameters', []);
-
-                return $handler->handle($request);
-            }
-        });
-
-        foreach ($middlewareStack as $middleware) {
-            $pipe->pipe($container->make($middleware));
-        }
-
-        $pipe->pipe(new HttpMiddleware\ExecuteRoute());
-
         $this->pipe = $pipe;
     }
